pyproject.toml
[project]
name = "git-scribe"
version = "0.1.0"
description = "Intelligence Layer between local code changes and remote repository. Automates commit messages and PR descriptions using LLMs."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Akash", email = "akash@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Version Control :: Git",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "typer>=0.9.0"
]

[project.scripts]
scribe = "git_scribe.main:app"

[project.urls]
Homepage = "https://github.com/example/git-scribe"
Repository = "https://github.com/example/git-scribe"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/git_scribe"]
